extractor-style refactor
