+++ background.html
@@ -1,0 +1,7 @@
+<!-- this is required for module importing -->
+<!DOCTYPE html>
+<html>
+<head>
+  <script type="module" src="background.js"></script>
+</head>
+</html>
