+++ docs/supportedsites.md
@@ -1,0 +1,76 @@
+# Supported Sites
+
+## Supported Handlers
+
+<table>
+<thead valign="bottom">
+<tr>
+    <th>Site</th>
+    <th>URL</th>
+    <th>Lexicons</th>
+    <th>NSID Resolution</th>
+</tr>
+</thead>
+<tbody valign="top">
+<tr>
+    <td>Bluesky</td>
+    <td>https://bsky.app</td>
+    <td>app.bsky.feed.post<br>app.bsky.graph.list<br>app.bsky.feed.generator<br>app.bsky.graph.starterpack</td>
+    <td>app.bsky</td>
+</tr>
+<tr>
+    <td>Langit</td>
+    <td>https://bsky.app</td>
+    <td>app.bsky.feed.post<br>app.bsky.graph.list<br>app.bsky.feed.generator<br>app.bsky.graph.starterpack</td>
+    <td></td>
+</tr>
+<tr>
+    <td>TOKEMI</td>
+    <td>https://tokimekibluesky.vercel.app<td>
+    <td>app.bsky.feed.post<br>app.bsky.graph.list<br>app.bsky.feed.generator<br>app.bsky.graph.starterpack</td>
+    <td></td>
+</tr>
+<tr>
+    <td>Aglais</td>
+    <td>https://aglais.pages.dev</td>
+    <td>app.bsky.feed.post<br>app.bsky.graph.list</td>
+    <td></td>
+</tr>
+<tr>
+    <td>Klearsky</td>
+    <td>https://klearsky.pages.dev</td>
+    <td>app.bsky.feed.post<br>app.bsky.graph.list<br>app.bsky.feed.generator<br>app.bsky.graph.starterpack</td>
+    <td></td>
+</tr>
+<tr>
+    <td>SkyChat</td>
+    <td>https://skychat.social</td>
+    <td>app.bsky.feed.post<br>app.bsky.graph.list<br>app.bsky.feed.generator<br>app.bsky.graph.starterpack</td>
+    <td></td>
+</tr>
+</tbody>
+</table>
+
+## Supported Resolvers
+
+<table>
+<thead valign="bottom">
+<tr>
+    <th>Domain Authority</th>
+    <th>Lexicons</th>
+    <th>URL Resolution</th>
+</tr>
+</thead>
+<tbody valign="top">
+<tr>
+    <td>app.bsky</td>
+    <td>app.bsky.actor.profile<br>app.bsky.feed.post<br>app.bsky.graph.list<br>app.bsky.feed.generator<br>app.bsky.graph.starterpack</td>
+    <td>https://bsky.app</td>
+</tr>
+<tr>
+    <td>XRPC</td>
+    <td>All URI</td>
+    <td>API calls are made directly to repo's PDS</td>
+</tr>
+</tbody>
+</table>
