+++ handler/atptools.js
@@ -1,0 +1,21 @@
+
+//   static atpTools = /^https:\/\/atp\.tools(?:\/at:\/(?:(?<handle>[\w.:%-]+)(?:\/(?<nsid>[\w.:%-]+)(?:\/(?<rkey>[\w.:%-]+)?)?)?)?)?(?:[?#].*)?$/
+// https://atp.tools/at:/ngerakines.me/blue.badge.collection
+import { getDid } from '../utils.js'
+
+export class ATPTools {
+    static DOMAINS = ['atp.tools']
+
+    static async processURL(url, settings, uriMode) {
+        const { handle, nsid, rkey } = this.parseURL(url)
+        console.log(`ATPTools handler recieved: ` + handle, nsid, rkey)
+        const did = await getDid(decodeURIComponent(handle))
+        if (!did) return null
+        return ['at:/', did, nsid, rkey].filter(Boolean).join("/")
+    }
+
+    static parseURL(url) {
+        let [handle, nsid, rkey] = url.pathname.split("/").slice(2)
+        return {handle, nsid, rkey}
+    }
+}
