+++ handler/bskycdn.js
@@ -1,0 +1,23 @@
+
+// static bskyCDN = /^https:\/\/cdn\.bsky\.app\/(?:[\w.:%-]+\/){3}(?<did>[\w.:%-]+)(?:\/[\w.:%-@]+)?$/
+// https://cdn.bsky.app/img/avatar/plain/did:plc:hx53snho72xoj7zqt5uice4u/bafkreiblp6bxadz42kevdcqcq3qofzxnoej5nhmwdqwcgjdg26ir2offdi@jpeg
+// static bskyVidCDN = /^https:\/\/video\.bsky\.app\/[\w.:%-]+\/(?<did>[\w.:%-]+)(?:\/[\w.:%-@]+)?$/
+// bskyVidCDN: https://video.bsky.app/watch/did%3Aplc%3Ahx53snho72xoj7zqt5uice4u/bafkreicf5nvgomy2kh3f6da7fcwghwrecm6ljcp4najy6vy5hvdn7tvec4/thumbnail.jpg
+
+// import { getDid } from '../utils.js'
+
+export class BskyCDN {
+    static DOMAINS = [ 'cdn.bsky.app', 'video.bsky.app' ]
+
+    static async processURL(url, settings, uriMode) {
+        const { did } = this.parseURL(url)
+        console.log(`BskyCDN handler recieved: ` + did)
+        return `at://${did}/blobs`
+    }
+
+    static parseURL(url) {
+        const sliceIndex = url.hostname === 'cdn.bsky.app' ? 4 : 2
+        const [did] = url.pathname.split("/").slice(sliceIndex)
+        return {did}
+    }
+}
