+++ handler/pdsls.js
@@ -1,0 +1,37 @@
+import { getDid } from '../utils.js'
+import { checkResolvers } from '../resolver/_resolvers.js'
+
+export class PDSls {
+    static DOMAINS = ['pdsls.dev']
+
+    static async processURL(url, settings, uriMode) {
+        const { pds, handle, nsid, rkey } = this.parseURL(url)
+        console.log(`PDSls handler recieved:`, pds, handle, nsid, rkey)
+
+        if (uriMode) {
+            const did = await getDid(handle)
+            if (!did) return null
+            return ['at:/', did, nsid, rkey].filter(Boolean).join('/')
+        }
+
+        if (settings.pdslsOpensApi) {
+            if (pds !== "at:") return `https://${pds}/xrpc/com.atproto.sync.listRepos?limit=1000`
+            const did = await getDid(handle)
+            if (!did) return null
+            return await Resolvers.xrpc({ did, nsid, rkey })
+        }
+
+        if (pds !== "at:") return `https://${pds}`
+        const did = await getDid(handle)
+        if (!did) return null
+        return await checkResolvers({ did, nsid, rkey }, settings)
+    }
+
+    static parseURL(url) {
+        let [pds, handle, nsid, rkey] = url.pathname.split('/').filter(Boolean)
+        return {pds, handle, nsid, rkey}
+    }
+}
+
+// static pdsls = /^https:\/\/pdsls\.dev\/(?<pds>[\w.%-]+)(?:\:\/)?(?:\/(?<handle>[\w.:%-]+))?(?:\/(?<nsid>[\w.:%-]+))?(?:\/(?<rkey>[\w.:%-]+))?(?:\/)?(?:[?#].*)?$/
+
