+++ resolver/_template.js
@@ -1,0 +1,7 @@
+export class  {
+    static NSID_AUTH =  ''
+
+    static async processURI({did, nsid, rkey}) {
+        
+    }
+}
